chore(executor): extract traits.rs marker traits (#116 step 8)
Pulls the tri-dialect marker trait family out of mod.rs into a
dedicated file (148 LOC):

- MaybePgFromRow / MaybeMyFromRow / MaybeSqliteFromRow — feature-gated
  FromRow<DbRow> bounds. Real bound when the feature is on; empty
  blanket-impl when off.
- LoadRelatedMy / LoadRelatedSqlite — per-dialect counterparts of
  LoadRelated. Macro-emitted for each Model so select_related joins
  stitch parents onto FK fields per backend.
- MaybeMyLoadRelated / MaybeSqliteLoadRelated — feature-gated wrappers
  for the above.

mod.rs: 2923 → 2791 LOC (-132). traits.rs: 148 LOC.

No public-API change. 1482 lib tests pass on sqlite,tenancy litmus.
